backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
sgp4==2.24
numpy==2.2.6
pandas==2.3.0
scikit-learn==1.7.0
scipy==1.15.3
requests==2.32.4
python-dateutil==2.9.0.post0
joblib==1.5.1
